wip command set db vars
